+++ azext_iot/iothub/common.py
@@ -83,6 +83,7 @@ IMMUTABLE_AND_DUPLICATE_MODULE_TWIN_FIELDS = [
 
 SYSTEM_ASSIGNED_IDENTITY = "[system]"
 BYTES_PER_MEGABYTE = 1048576
+NON_DECODABLE_PAYLOAD = "{{non-decodable payload}}"
 
 # Message Endpoint Messages
 INVALID_CLI_CORE_FOR_COSMOS = "This version of the azure cli core does not support Cosmos Db Endpoints for IoT Hub."